refactored existing tests
