+++ spec/features/comments_feature_spec.rb
@@ -1,0 +1,5 @@
+require "rails_helper"
+
+feature "comments" do
+  
+end
